feat: optimize dependencies and composition of classes
