+++ src/common/index.ts
@@ -1,3 +1,3 @@
-export * from './declare';
 export * from './interface';
+export * from './declare';
 export * from './type';
