+++ src/struct/apollo-client-builder.ts
@@ -1,51 +1,24 @@
-import { ApolloClientBuildOptions, IApolloConfigMapper } from '../common';
+import { ApolloClientBuildOptions } from '../common';
 
-import { ApolloClient } from './apollo-client';
-import { ApolloConfigLoader } from './apollo-config-loader';
+import { ApolloConfigClient } from './apollo-client';
 import { ApolloConfigMapper } from './apollo-config-mapper';
+import { ApolloConfigObserver } from './apollo-config-observer';
 import { ApolloHttpClient } from './apollo-http-client';
 
 export class ApolloClientBuilder {
-  private client: ApolloClient;
+  public build(options: ApolloClientBuildOptions) {
+    const { clientOptions, clientMapper, clientPollRetry, clientPollDelay } = options;
 
-  private loader: ApolloConfigLoader;
+    const mapper = clientMapper || new ApolloConfigMapper();
+    const httpClient = new ApolloHttpClient(clientOptions);
+    const observer = new ApolloConfigObserver(httpClient);
 
-  private mapper: IApolloConfigMapper;
-
-  private httpClient: ApolloHttpClient;
-
-  constructor(
-    private readonly options: ApolloClientBuildOptions,
-  ) {
-    this.httpClient = new ApolloHttpClient(this.options.clientOptions);
-    this.mapper = this.options.configMapper || new ApolloConfigMapper();
-  }
-
-  public buildLoader() {
-    const { configEnv, configFilePath, configFileType } = this.options;
-
-    this.loader = new ApolloConfigLoader(
-      configFilePath,
-      configEnv,
-      configFileType,
-      this.httpClient,
-      this.mapper,
+    return new ApolloConfigClient(
+      httpClient,
+      mapper,
+      observer,
+      clientPollRetry,
+      clientPollDelay,
     );
-
-    return this;
-  }
-
-  public buildClient() {
-    const { clientPollRetry, clientPollDelay } = this.options;
-
-    this.client = new ApolloClient(this.httpClient, this.mapper, clientPollRetry, clientPollDelay);
-    return this;
-  }
-
-  public export() {
-    return {
-      client: this.client,
-      loader: this.loader,
-    };
   }
 }
