+++ src/struct/index.ts
@@ -1,6 +1,7 @@
 export * from './apollo-client-builder';
 export * from './apollo-client';
-export * from './apollo-config-loader';
 export * from './apollo-config-mapper';
-export * from './apollo-config-signer';
+export * from './apollo-config-observer';
+export * from './apollo-http-client-signer';
 export * from './apollo-http-client';
+export * from './apollo-third-party-http-client';
