--- test/files/json/DEFAULT.json
@@ -1,24 +1,0 @@
-{
-  "apollo": {
-    "domain": {
-      "listenNamespace2": {
-        "options": {
-          "namespace": "listenNamespace2",
-          "type": "properties"
-        }
-      },
-      "listenNamespace3": {
-        "options": {
-          "namespace": "listenNamespace3",
-          "type": "json"
-        }
-      },
-      "listenNamespace4": {
-        "options": {
-          "namespace": "listenNamespace3",
-          "type": "json"
-        }
-      }
-    }
-  }
-}
