--- test/files/json/DEV.json
@@ -1,12 +1,0 @@
-{
-  "apollo": {
-    "domain": {
-      "client": {
-        "options": {
-          "namespace": "client",
-          "type": "json"
-        }
-      }
-    }
-  }
-}
