--- test/files/yml/DEFAULT.yml
@@ -1,10 +1,0 @@
-apollo:
-  domain:
-    listenNamespace2:
-      options:
-        namespace: "listenNamespace2"
-        type: "properties"
-    listenNamespace3:
-      options:
-        namespace: "listenNamespace3"
-        type: "json"
